refactor: move `processOrderMessage` method directly to `OrdersWorkerService`
